+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/roomlist/room/ArchivedRoom.kt
@@ -26,8 +26,9 @@ import de.connect2x.messenger.compose.view.theme.components.ModalDialogFooter
 import de.connect2x.messenger.compose.view.theme.components.ModalDialogHeader
 import de.connect2x.messenger.compose.view.theme.components.ThemedButton
 import de.connect2x.messenger.compose.view.theme.components.ThemedIconButton
+import de.connect2x.messenger.compose.view.theme.components.ThemedInfoChip
+import de.connect2x.messenger.compose.view.theme.components.ThemedLabel
 import de.connect2x.messenger.compose.view.theme.components.ThemedModalDialog
-import de.connect2x.messenger.compose.view.util.TextLabel
 import de.connect2x.trixnity.messenger.viewmodel.roomlist.RoomListElementViewModel
 
 interface ArchivedRoomListElement {
@@ -50,7 +51,7 @@ class ArchivedRoomListElementImpl : ArchivedRoomListElement {
             roomListElementViewModel = roomListElementViewModel,
             extraInfo = {
                 Spacer(Modifier.size(5.dp))
-                TextLabel(i18n.commonArchived())
+                ThemedLabel(i18n.commonArchived())
             }
         ) {
             Tooltip(